Door Method
Added a door monobehaviour to attach to doors. Relies on the BoxCollider2D to open and close the door (Can be changed if needed). Has a public function called ChangeState() which toggles the doors state (if it is open or closed). Then runs a function depending on its new state.