Default Respawn Function + WIP Changeable Respawn Implementation
Added Respawn() function to Player.cs which by default will respawn the player at (0, 0) and allows for its position to be changed by the RespawnPoint object saved in it.

Added a new RespawnPoint prefab which contains RespawnPoint.cs which stores the respawn position. It implements IInteractable (it has 2 i's not sure if spelling error).

WIP - Currently missing implementation where when running the Interact() function in RespawnPoint.cs, it would then change the CurrentRespawnPoint variable in Player.cs