Fixed FMOD and added debug toggles / test audio for sliders
Accidently botched FMOD so audio would not play so fixed in this one. Debug toggle has been added as well. Forgot to mention that Persistent object which contains settings is added to TestScene_GameplayUI for well... testing.

How to implement volume sliders:
1. Drag the "Persistent" prefab into the scene
2. Select the child object "Settings" under "Persistent"
3. Drag the UI sliders into each slot
4. To test, ensure that "Enable Debug" and "Audio Debug" are on in "Settings" to allow for test audio to be played.

Note: There is a slider for each volume built into the inspector but it only works if the settings for the corresponding slider are null to prevent conflicts with actual slider.