Committed YamlDotNet library used for parsing
Note: I left the .meta files in the commit because we will likely never
need to recompile this. I'm hoping this saves compile time. I also did
not try to import only a subset of the library, so this might be too
much, but it was safer to do it this way.